Created separate component for PhoneAuth
